krige_p_pvals

Creates a kriging plot from ST data with xCell pvalues.

krige_p_pvals(
  data_f = NULL,
  mask = NULL,
  color_pal = "YlOrBr",
  leg_name = "",
  title_name = "",
  x = NULL,
  plot_who = NULL,
  cell = NULL
)

Arguments

data_f,
a data with three columns: x coordinates, y coordinates, and the kriging prediction values to be plotted.

mask,
an object of class SpatialPolygons containing a large polygon encasing all the predicted grid, and a smaller polygon drawing the concave hull of the tissue shape.

color_pal,
a scheme from 'khroma'. Defaul is 'YlOrBr'

leg_name,
a short name for the legend title.

title_name,
a short name for the plot title.

x,
an STList from where coordinates will be taken.

plot_who,
an integer indicating the spatial array to be plotted.

cell,
a cell name from a deconvoluted matrix.

Value

, a ggplot object.

Details

Function to produce a "kriging plot" from a data frame with three columns: x coordinates, y coordinates, and predicted kriging values. The data frame has column names 'x_pos', 'y_pos', and 'krige'. The function also takes a SpatialPolygons object to mask the predicted grid to the area of the tissue. It also takes a color palette name from the 'khroma' package. Finally, it takes a name for the color legend title. Finally, the function overlays dots for spots in which significant p-values (<0.05) were observed for a cell type.